Created TaskList and updated Task
Created the TaskList class, which will be used to maintain the list of Tasks as they are entered by the user. Updated Task to implement the Comparable interface, making it easier to compare Tasks as they are added to TaskList. Added a few short checkpoint tests to the main method of TaskManager. The next step is to incorporate the TaskList functionality with the Task Manager GUI.