Began Working on GUI Application
Completed several features of the GUI Application including some key labels, being able to enter new tasks, entering new tasks, and updating the display of tasks when new ones are entered. The next step is to add a way to remove tasks, and then to incorporate File I/O so that results can be stored/read between uses.